% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summaries.R
\name{char_summary}
\alias{char_summary}
\title{Summary for character vector}
\usage{
char_summary(x, na.rm = FALSE)
}
\arguments{
\item{x}{character vector}

\item{na.rm}{whether to remove NAs}
}
\value{
a named list of 3
}
\description{
Description blah blah
}
\examples{

x <- c("Boston", "NY", "Boston", "New Haven")
char_summary(x)

}
