% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalulate_signatures.R
\name{GSVAsignatureRanking}
\alias{GSVAsignatureRanking}
\title{This function evaluates the contributions of genes in GSVA scores}
\usage{
GSVAsignatureRanking(eset, signature)
}
\arguments{
\item{eset}{an expression set}

\item{signature}{a list of signatures (at least 1)}
}
\value{
a list of a dataframe and two heatmaps
}
\description{
This function evaluates the contributions of genes in GSVA scores
}
