% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_summaries.R
\name{numeric_summary}
\alias{numeric_summary}
\title{Summary for a numeric vector}
\usage{
numeric_summary(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector}

\item{na.rm}{whether to include NA values in the calculations}
}
\value{
a named numeric vector with 6 values
}
\description{
\code{numeric_summary} calculates extended summary for a numeric vector
}
\examples{

x <- rnorm(100)
numeric_summary(x)

}
