% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_signature_heatmap.R
\name{omics_signature_heatmap}
\alias{omics_signature_heatmap}
\title{This function generates an annotated heatmap that shows the contributions of genes to the given signature score}
\usage{
omics_signature_heatmap(
  eset,
  signature,
  sig_score = NULL,
  col_ha = NULL,
  method = c("GSVA"),
  ...
)
}
\arguments{
\item{eset}{an expression set}

\item{signature}{a list of signatures (at least 1)}

\item{sig_score}{an aggregate signature score}

\item{col_ha}{a ComplexHeatmap::heatmapAnnotation object with columns' (i.e., samples') annotation}

\item{method}{how to compute the aggregate score (only GSVA at the moment)}

\item{...}{additional parameters to pass to the method}
}
\value{
a list of a dataframe and two heatmaps
}
\description{
This function generates an annotated heatmap that shows the contributions of genes to the given signature score
}
