% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omics_signature_score.R
\name{omics_signature_score}
\alias{omics_signature_score}
\title{This function computes an aggregate score for a set of genes}
\usage{
omics_signature_score(eset, signature, method = c("GSVA"), ...)
}
\arguments{
\item{eset}{an expression set}

\item{signature}{a list of signatures (at least 1)}

\item{method}{how to compute the aggregate score (only GSVA at the moment)}

\item{...}{additional parameters to pass to the method}
}
\value{
a named numeric vector of scores (w/ the names corresponding to colnames(eset))
}
\description{
This function computes an aggregate score for a set of genes
}
